shell: getParameter() em ViewData é parametrizável.
Sem necessidade para casting na chamada.

+++ datadict/classes/org/iocaste/datadict/Common.java
@@ -87,7 +87,7 @@ public class Common {
      * @return
      */
     public static final byte getMode(ViewData vdata) {
-        return (Byte)vdata.getParameter("mode");
+        return vdata.getParameter("mode");
     }
     
     /**
